runtime: Introduce bipolarity in conditions
Introduce bipolarity in condition merge options to allow bipolar
conditions to behave in expected ways.

Unlike the normal-false or abnormal-true pattern of conditions,
BiPolar conditions are the opposite, they are normal-true and
abnormal-false. But when they are abnormal/false, they have to be
considered like any negative polarity condition.

Since BiPolar conditions are always present regardless of their value,
they need to have the higher priority than the typical negative
polarity as absense of negative polarity but presence of false bipolar
condition is a failure case. Due to highest priority, BiPolar
conditions must be carefully used. A False value would result in a
false result.

This change shifts the merge priority score by one point and positions
the BiPolarity=False to have the highest merge priority score.
New ConditionGroup based on bipolarity value are created while
processing the conditions during merge.

This is implemented to be orthogonal to the existing behavior when not
using bipolarity conditions in merge.

Signed-off-by: Sunny <darkowlzz@protonmail.com>